Improve smaug support (#36)
* Remove assumption about position of zif files in project.

`require.rb` currently assumes that the `zif` project is added into
an `app/lib/zif` folder. This means you need to explicitly require
each file individually if that assumption is broken (e.g. when zif is
installed via Smaug).

By using `require_relative` instead, `require.rb` can now be used
regardless of where the zif library is placed.

* Simplify smaug package.

Currently, when zif is installed through smaug, the whole repository
is copied over, including docs, the jekyll site, linters config and the
whole example app.

With this change, we instruct smaug to only copy the library files,
the README, the license and the smaug config.